+++ packages/components/dashboard/index.ts
@@ -1,0 +1,6 @@
+export * from './module';
+export * from './dashboard.component';
+export * from './widget/body/widget-body.component';
+export * from './widget/header/widget-header.component';
+export * from './widget/widget.component';
+export * from './dashboard.class';